requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
idna==3.4
psycopg2==2.9.3
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
